pyproject.toml
[tool.poetry]
name = "nonebot-plugin-drawer"
version = "0.2.7"
description = "适用于 Nonebot2 的AI画画插件(对接文心大模型API)"
authors = ["CrazyBoyM <ai-lab@foxmail.com>"]
license = "MIT License"
readme = "README.md"
homepage = "https://github.com/CrazyBoyM/nonebot-plugin-drawer"
repository = "https://github.com/CrazyBoyM/nonebot-plugin-drawer"

[tool.poetry.dependencies]
python = "^3.7"
httpx = "^0.19.0"
nonebot2 = "^2.0.0-beta.1"
pydantic = "^1.9.2"
nonebot-adapter-onebot = "^2.1.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
